+++ stories/DesignSystemOverview.stories.ts
@@ -1,0 +1,11 @@
+import DesignSystemOverview from './DesignSystemOverview.svelte';
+
+export default {
+  title: 'Design System/Overview',
+  component: DesignSystemOverview,
+  parameters: {
+    layout: 'fullscreen'
+  }
+};
+
+export const Overview = {};